FEAT: create produtoDTO

+++ src/main/java/com/jogos/ecommerce/domain/model/Cargo.java
@@ -30,7 +30,7 @@ public class Cargo {
     private Long id;
     
     @NotBlank
-    private float salario;
+    private double salario;
     @NotBlank
     private  String descricao;
